Move addToWishlist command in connectors
