+++ web/app/connectors/_sfcc-connector/cart/commands.js
@@ -2,7 +2,7 @@
 /* Copyright (c) 2017 Mobify Research & Development Inc. All rights reserved. */
 /* * *  *  * *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  * */
 
-import {makeApiRequest, makeApiJsonRequest, getAuthTokenPayload, checkForResponseFault} from '../utils'
+import {makeApiRequest, makeApiJsonRequest} from '../utils'
 import {populateLocationsData} from '../checkout/utils'
 import {requestCartData, createBasket, handleCartData, createNewBasket, isCartExpired, updateExpiredCart} from './utils'
 
@@ -70,45 +70,5 @@ export const updateCartItem = (itemId, quantity, productId) => (dispatch) => {
 export const updateItemQuantity = (itemId, quantity) => (dispatch) => dispatch(updateCartItem(itemId, quantity))
 
 export const initCartPage = () => (dispatch) => Promise.resolve(dispatch(populateLocationsData()))
-
-const NEW_WISHILIST_PAYLOAD = {
-    type: 'wish_list',
-    name: 'Saved for Later'
-}
-
-export const addToWishlist = (productId) => (dispatch) => {
-    const {sub} = getAuthTokenPayload()
-    const customerID = JSON.parse(sub).customer_info.customer_id
-
-    return makeApiRequest(`/customers/${customerID}/product_lists`, {method: 'GET'})
-        .then((response) => response.json())
-        .then(({count, data}) => {
-            if (count) {
-                return data[0]
-            }
-            // create a list if one doesn't exist
-            return makeApiJsonRequest(
-                `/customers/${customerID}/product_lists`,
-                NEW_WISHILIST_PAYLOAD,
-                {method: 'POST'}
-            )
-            .then(checkForResponseFault)
-        })
-        .then(({id}) => {
-            const requestBody = {
-                type: 'product',
-                product_id: productId,
-                quantity: 1
-            }
-
-            return makeApiJsonRequest(
-                `/customers/${customerID}/product_lists/${id}/items`,
-                requestBody,
-                {method: 'POST'}
-            )
-            .then(checkForResponseFault)
-        })
-        .catch(() => { throw new Error('Unable to add item to wishlist.') })
-}
 
 export const fetchTaxEstimate = () => Promise.reject('Method not implemented')
